+++ test/test_transforms.py
@@ -2,7 +2,7 @@ import os
 import torch
 import torchvision.transforms as transforms
 import torchvision.transforms.functional as F
-import torchvision.transforms.functional_tensor as F_t
+import torchvision.transforms._functional_tensor as F_t
 from torch._utils_internal import get_file_path_2
 from numpy.testing import assert_array_almost_equal
 import unittest
